Split local repo into its own module
