+++ owners/src/repo/repo.js
@@ -14,27 +14,6 @@
  * limitations under the License.
  */
 
-const path = require('path');
-const fs = require('fs');
-const childProcess = require('child_process');
-const util = require('util');
-const exec = util.promisify(childProcess.exec);
-
-/**
- * Execute raw shell commands.
- *
- * @private
- * @param {...string} commands list of commands to execute.
- * @return {!Array<string>} command output
- */
-async function runCommands(...commands) {
-  return exec(commands.join(' && '))
-    .then(({stdout, stderr}) => stdout)
-    .catch(({stdout, stderr}) => {
-      throw stderr;
-    });
-}
-
 /**
  * Interface for reading from a GitHub repository.
  */
@@ -66,97 +45,4 @@ class Repository {
   }
 }
 
-/**
- * Interface for reading from a checked out repository using relative paths.
- */
-class LocalRepository extends Repository {
-  /**
-   * Constructor.
-   *
-   * @param {string} pathToRepoDir absolute path to the repository root
-   *     directory.
-   * @param {?string=} remote git remote to clone (default: 'origin').
-   */
-  constructor(pathToRepoDir, remote) {
-    super();
-    this.rootDir = pathToRepoDir;
-    this.remote = remote || 'origin';
-  }
-
-  /**
-   * Checks out the master branch.
-   */
-  async sync() {
-    await this.checkout();
-  }
-
-  /**
-   * Runs commands in the repository's root directory.
-   *
-   * @param {...string} commands list of commands to execute.
-   * @return {!Array<string>} command output
-   */
-  async _runCommands(...commands) {
-    return await runCommands(`cd ${this.rootDir}`, ...commands);
-  }
-
-  /**
-   * Check out a branch locally.
-   *
-   * @param {?string=} [branch=master] git branch to checkout.
-   */
-  async checkout(branch) {
-    branch = branch || 'master';
-    await this._runCommands(
-      `git fetch ${this.remote} ${branch}`,
-      `git checkout -B ${branch} ${this.remote}/${branch}`
-    );
-  }
-
-  /**
-   * Get an absolute path for a relative repo path.
-   *
-   * @param {string} relativePath file or directory path relative to the root.
-   * @return {string} absolute path to the file in the checked out repo.
-   */
-  _getAbsolutePath(relativePath) {
-    return path.resolve(this.rootDir, relativePath);
-  }
-
-  /**
-   * Read the contents of a file from the checked out repo.
-   *
-   * @param {string} relativePath file to read.
-   * @return {string} file contents.
-   */
-  async readFile(relativePath) {
-    const filePath = this._getAbsolutePath(relativePath);
-    return fs.readFileSync(filePath, {encoding: 'utf8'});
-  }
-
-  /**
-   * Finds all OWNERS files in the checked out repository.
-   *
-   * Assumes repo is already checked out.
-   *
-   * @return {!Array<string>} a list of relative OWNERS file paths.
-   */
-  async findOwnersFiles() {
-    // NOTE: for some reason `git ls-tree --full-tree -r HEAD **/OWNERS*`
-    // doesn't work from here.
-    const ownersFiles = await this._runCommands(
-      [
-        // Lists all files in the repo with extra metadata.
-        'git ls-tree --full-tree -r HEAD',
-        // Cuts out the first three columns.
-        'cut -f2',
-        // Finds OWNERS files.
-        'egrep "OWNERS$"',
-      ].join('|')
-    );
-
-    return ownersFiles.trim().split('\n');
-  }
-}
-
-module.exports = {Repository, LocalRepository};
+module.exports = {Repository};
