+++ owners/test/index.test.js
@@ -19,7 +19,7 @@ const nock = require('nock');
 const owners = require('..');
 const {Probot} = require('probot');
 const sinon = require('sinon');
-const {LocalRepository} = require('../src/repo/repo');
+const LocalRepository = require('../src/repo/local_repo');
 const {GitHub, Team} = require('../src/github');
 const {OwnersParser} = require('../src/parser');
 const {UserOwner} = require('../src/owner');
